Remove inconsistent comparison for Keys, overhaul matching of Bindings
Before this commit, registering multiple providers for the same type but with different classifiers would  fail, because all but the first Provider were not registered in the BindingList. Key and therefore BindingEntry implemented compareTo() inconsistent with equals(); the TreeSet baking the BindingList would therefore recognise providers of the same type but with different classifiers as duplicates.
As of this commit, Key and BindingList now no longer implement Comparable. BindingList uses a simple HashMultimap.

Before this commit, Key and BindingEntry did not implement equals() and hashCode(). To compare two Keys, a Key.match(Key)  was used; it returned a positive if both Keys had the same Type and if the instance match() was invoked on had a classifier, the given instance had the same classifier. Therefore two keys where also considered as match, if the first Key did not have a classifier, but the second did (this seems to be unintended behavior).
As of this commit, the match() method is removed. Instead Key and BindingEntry implement equals() and hashCode(): Two Keys are equal if and only if they have the same type and either both none or the same classifier.

Before this commit, multiple Providers could be registered for the same Key (same Type and classifier); BindingList.addBinding() would fail quietly.
As of this commit, addBinding() throws an IllegalArgumentException if a Provider is already registered for a Key that is equal to the given one.